fix(gha,benchmark): GITHUB_SHA isn't the sha that triggered pr event.
When the event is of type `pull_request`, the `GITHUB_SHA` environment
variable contains the sha of the merge branch created under the hood.
The merge commit is not reachable in the project's git tree.

It was changed to use `${{ github.event.pull_request.head.sha }}` which
points to the head of the PR at the time that the event triggered.
Note: when re-running the event, the same payload is used then the same
HEAD commit is used even if PR HEAD moved.

Signed-off-by: Tiago Natel <t.nateldemoura@gmail.com>