chore(tmc): use real terraform in e2e tests.
## Reasons for this change

Before we used a fake terraform binary that mimicked some Terraform behavior and it supported fully customization of the program's output.
The problem is that the fake terraform misses a lot of important details:
- `terraform show` gives different error messages depending on different requirements not met.
- `terraform show` requires an initialized terraform module and then mimicking this behavior is hard.
- `terraform show` checks the lock files
- etc

if we ignore those details, we miss the opportunity to test those cases and improve Terramate error messages for such errors.
_The error message improvements are not addressed in this PR_.

## Changes

- Now during the e2e test setup, the Terraform is installed in a temp directory using the hashicorp official `hc-install` library.
- Tests were updated to use a real project.
- Sandbox library updated to support `copy` and `run` commands.