chore(tests): use detected Terraform if available.
Add support for running e2e tests with a different version of
Terraform when exporting the `TM_TEST_TERRAFORM_REQUIRED_VERSION`
environment variable.

Additionally, the tests now detect if Terraform is already
installed and then the detected version is used. If the detected
version does not match the version provided in the environment
variable, then it's installed as before.

Signed-off-by: Tiago Natel <t.nateldemoura@gmail.com>