feat(tmc): send logs of the user-supplied command to Terramate Cloud.
# Reasons for This Change

Terramate Cloud supports visualizing the deployment logs in real-time.

## Description of the changes

A log syncer was created that proxies stdout/stderr writes using an in-memory pipe. The syncer creates 3 goroutines (1 for stdout, 1 for stderr and 1 for synchronizing logs) that *MUST* be shutdown/terminated after execution of each stack.
The syncer supports a customizable `batchSize` and `maxLineSize` but the requirement of the latter will be fixed in a separate PR.

Signed-off-by: Tiago Natel <t.nateldemoura@gmail.com>