fix(generate): tm_dynamic attributes over wrapped.
The attributes node were being wrapped with a CloneExpression node type
over and over at each tm_dynamic generation for no reason, leading to
stack exhaustion in the recursive ast.CloneExpr() function.
The node needs to be wrapped because PartialEval does in-place
substitution then it needs to be cloned if used multiple times.

Signed-off-by: Tiago Natel <t.nateldemoura@gmail.com>