chore: fix --cloud-status tests.
The tests were using a test environment variable for injecting
the deployment UUID for allowing the retrieval of deployments
assoaciated with a given testcase. The problem is that doing that
we cannot test a cli code path that assigns the cli cloud context
(`cli.cloud.run.runUUID`). This problem led to the issue fixed
by #1484.

Signed-off-by: Tiago Natel <t.nateldemoura@gmail.com>