chore: Check goreleaser version and update Go Releaser to v1.14.0-pro
This commit updates the Go Releaser version in the common makefile to v1.14.0-pro. It also adds a new target, `check-goreleaser-version`, which checks the installed version of Go Releaser against the specified version. If the installed version does not match, an error is thrown. This ensures that the correct version of Go Releaser is being used for the release process.