fix(parallel): deadlock in case command does not exist. (#1832)
Note: this is just a quick fix and a refactor in this code path is being
worked on.

This fixes a missing `releaseResource()` that led to a deadlock in the
case that user-supplied command does not exist. The fix also release the
resource in all other code paths that abort the execution, for
correctness purposes, but AFAICS they are optional.

Relates to #1828

A refactor in the outputs-sharing handling is being designed.

```
yes, fixes a --parallel bug.
```